+++ tests/test_gait_analysis.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from paradigma.gait.gait_analysis import filter_gait_io, detect_gait_io, extract_arm_activity_features_io, extract_gait_features_io
 from paradigma.config import FilteringGaitConfig, ArmActivityFeatureExtractionConfig, ArmSwingQuantificationConfig, GaitDetectionConfig, GaitFeatureExtractionConfig
-from paradigma.imu_preprocessing import preprocess_imu_data_io
+from paradigma.preprocessing import preprocess_imu_data_io
 from test_notebooks import compare_data
