+++ tests/test_ppg_analysis.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from paradigma.heart_rate.heart_rate_analysis import extract_signal_quality_features
 from paradigma.config import PPGConfig, IMUConfig, SignalQualityFeatureExtractionConfig
-from paradigma.ppg_preprocessing import preprocess_ppg_data, scan_and_sync_segments
+from paradigma.preprocessing import preprocess_ppg_data, scan_and_sync_segments
 from test_notebooks import compare_data
